refactor: introduce http-utils
